Search and Sort - Hooks + Tests

+++ client/src/app/components/home/home.component.ts
@@ -14,6 +14,22 @@ import { supportsPassiveEventListeners } from '@angular/cdk/platform';
 
 
 
+function dynamicSort(property) {
+  var sortOrder = 1;
+
+  if(property[0] === "-") {
+      sortOrder = -1;
+      property = property.substr(1);
+  }
+
+  return function (a,b) {
+      if(sortOrder == -1){
+          return b[property].localeCompare(a[property]);
+      }else{
+          return a[property].localeCompare(b[property]);
+      }        
+  }
+}
 
 @Component({
   selector: 'app-home',
@@ -28,6 +44,16 @@ export class HomeComponent implements OnInit {
   displayedItems = [];
   destroy$: Subject<boolean> = new Subject<boolean>();
   editingItemIndex: Number;
+
+  searchAndSortForm: FormGroup;
+
+  searchBar: FormControl;
+  sortBy: FormControl;
+  previousSortCriteria: string;
+  sortAscending = true;
+
+  previousWishListItems: FormArray;
+
   
   itemFormGroup = {
     'title': [null, Validators.required],
@@ -99,6 +125,9 @@ export class HomeComponent implements OnInit {
   constructor(private apiService: ApiService, private formBuilder: FormBuilder, public dialog: MatDialog) { 
    
     this.editingItemIndex = -1;
+    this.searchBar = new FormControl('');
+    this.sortBy = new FormControl('add_date');
+    this.previousSortCriteria = this.sortBy.value;
   }
 
 
@@ -147,7 +176,37 @@ export class HomeComponent implements OnInit {
       // wishlist_items: this.formBuilder.array([this.formBuilder.group(this.itemFormGroup)])
     })
 
+    this.searchAndSortForm = this.formBuilder.group({ 
+      searchBar: [null],
+      sortBy: [null] 
+    });
+
+    this.previousWishListItems = this.formBuilder.array([]);
+
+    this.onSearchBarValueChanges();
+    this.onSortByValueChanges();
     this.refreshData();
+    
+  }
+
+  onSearchBarValueChanges(): void {
+    this.searchBar.valueChanges.subscribe(val => {
+      this.filterItems(val);
+    });
+  }
+
+  onSortByValueChanges(): void {
+    this.sortBy.valueChanges.subscribe(val => {
+      
+      this.sortItemsBy(val);
+            
+    });
+  }
+
+  onSortByClick() {
+    
+    this.sortAscending = !this.sortAscending;
+    this.sortItemsBy(this.sortBy.value);
     
   }
 
@@ -155,18 +214,11 @@ export class HomeComponent implements OnInit {
     this.isLoading = true;
     this.apiService.sendGetRequest().pipe(takeUntil(this.destroy$)).subscribe((res: HttpResponse<Array<Item>>)=>{  
       
-      this.items = res.body;
-      
-      while (this.wishlistItems.length !== 0) {
-        this.wishlistItems.removeAt(0)
-      }
-      this.items.forEach(element => {
-        this.wishlistItems.push(this.formBuilder.group(element));
-      });
-      this.addItem();
+      this.fillFormArrayWithArray(this.wishlistItems, res.body);
       this.editingItemIndex = this.wishlistItems.length - 1;
-      // console.log(res.body);  
-      this.displayedItems = res.body;
+
+      this.fillFormArrayWithArray(this.previousWishListItems, res.body);
+      
       this.isLoading = false;
     })
   }
@@ -199,8 +251,8 @@ export class HomeComponent implements OnInit {
     });
 
     dialogRef.afterClosed().subscribe(dialogResult => {
+      if (dialogResult == false) return;
       this.isLoading = true;
-      // FIXME (add check on result, otherwise I'll delete in each case :)
       this.apiService.deleteItemById(item.id)
         .subscribe(res => {
             //let id = res['id'];
@@ -210,11 +262,65 @@ export class HomeComponent implements OnInit {
           }, (err) => {
             console.log(err);
             this.isLoading = false;
-          });
-          
+          });         
     });
   }
 
+  fillFormArrayWithArray(formArray: FormArray, array: any[]): FormArray {
+    formArray.clear();
+    
+    for (let i=0;i<array.length;i++) {
+      formArray.push(this.formBuilder.group(array[i]));
+    }
+    formArray.push(this.formBuilder.group(this.itemFormGroup));
+    return formArray;
+  }
+
+  
+
+  filterItems(val: string) {
+
+    let _displayedItemes = [];
+    this.previousWishListItems.value.forEach(element => {
+      if(element.id && element.id > 0) _displayedItemes.push(element);
+    });
+    
+    let _wishlistItemsArr = _displayedItemes.filter(item => {
+      return (
+        item.title.toLocaleLowerCase().indexOf(val.toLocaleLowerCase()) !== -1 ||
+        item.description.toLocaleLowerCase().indexOf(val.toLocaleLowerCase()) !== -1 
+      );
+    });
+    this.fillFormArrayWithArray(this.wishlistItems, _wishlistItemsArr);  
+    this.editingItemIndex = this.wishlistItems.length - 1;
+
+  }
+
+  sortItemsBy(val: string) {
+    
+    let _displayedItemes = [];
+    this.wishlistItems.value.forEach(element => {
+      if(element.id && element.id > 0) _displayedItemes.push(element);
+    });
+    if (val == 'name') {
+      _displayedItemes = _displayedItemes.sort(dynamicSort("title"));
+    }
+    else {
+      //console.log(typeof this.displayedItems);
+      _displayedItemes = _displayedItemes.sort(function(a, b){
+        let keyA = new Date(a.add_date),
+            keyB = new Date(b.add_date);
+        // Compare the 2 dates
+        if(keyA < keyB) return -1;
+        if(keyA > keyB) return 1;
+        return 0;
+      });
+    }
+    if(!this.sortAscending) _displayedItemes = _displayedItemes.reverse();
+    
+    this.fillFormArrayWithArray(this.wishlistItems, _displayedItemes);
+    this.previousSortCriteria = val;
+  }
